+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Add RouteResolver to resolve platform.sh routes into their environment URLs
 
 ## [1.0.2] - 2017-09-19
 ### Changed